Refactor tab appearance.
Rename CSS module.
Make tabs appear more like browser tabs.
Use CSS module more effectively.
Add placeholder for the case that no file is loaded in DynamicTab.
