--- src/ui/widgets/Tabs/navigationTabs.module.css
@@ -1,12 +1,0 @@
-.Bar {
-  display: flex;
-  overflow: auto;
-  justify-content: flex-start;
-}
-
-.Button {
-  margin: 0.2em;
-  font-size: 2rem;
-  font-weight: bold;
-  min-width: 20%;
-}
