+++ src/ui/widgets/Tabs/tabContainer.tsx
@@ -15,7 +15,7 @@ import { errorWidget } from "../EmbeddedDisplay/embeddedDisplay";
 import { parseJson } from "../EmbeddedDisplay/jsonParser";
 import { widgetDescriptionToComponent } from "../createComponent";
 
-import classes from "./navigationTabs.module.css";
+import classes from "./tabs.module.css";
 
 export const TabContainerProps = {
   tabs: PropTypes.objectOf(PropTypes.object).isRequired,
@@ -53,15 +53,18 @@ export const TabContainerComponent = (
 
   return (
     <div>
-      <div className={classes.Bar}>
+      <div className={classes.TabBar}>
         {Object.keys(props.tabs).map(
           (key, index): JSX.Element => (
             <button
               onClick={(): void => {
                 setIndex(index);
               }}
-              className={classes.Button}
-              style={index === childIndex ? { borderStyle: "inset" } : {}}
+              className={
+                index === childIndex
+                  ? `${classes.Tab} ${classes.TabSelected}`
+                  : classes.Tab
+              }
               key={index}
             >
               {key}
